Move runService method into routerlicious
