+++ server/routerlicious/packages/routerlicious/src/alfred/www.ts
@@ -5,15 +5,15 @@
 
 import * as path from "path";
 import * as winston from "winston";
-import * as core from "@fluidframework/server-services-core";
 import { configureLogging } from "@fluidframework/server-services-utils";
 import { AlfredResourcesFactory, AlfredRunnerFactory } from "@fluidframework/server-routerlicious-base";
+import { runService } from "../runner";
 
 const configPath = path.join(__dirname, "../../config/config.json");
 
 configureLogging(configPath);
 
-core.runService(
+runService(
     new AlfredResourcesFactory(),
     new AlfredRunnerFactory(),
     winston,
