+++ server/routerlicious/packages/routerlicious/src/riddler/www.ts
@@ -5,15 +5,15 @@
 
 import * as path from "path";
 import * as winston from "winston";
-import * as core from "@fluidframework/server-services-core";
 import { configureLogging } from "@fluidframework/server-services-utils";
 import { RiddlerResourcesFactory, RiddlerRunnerFactory } from "@fluidframework/server-routerlicious-base";
+import { runService } from "../runner";
 
 const configPath = path.join(__dirname, "../../config/config.json");
 
 configureLogging(configPath);
 
-core.runService(
+runService(
     new RiddlerResourcesFactory(),
     new RiddlerRunnerFactory(),
     winston,
